requirements.txt
semantic-kernel
google-generativeai
serpapi
newspaper3k
